requirements.txt
gymnasium>=0.29
torch>=2.0
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
imageio>=2.31
tqdm>=4.65
scipy>=1.11
wandb>=0.16
seaborn>=0.12
pandas>=2.0
